Moved hdf functionality to anesthetic.read
