+++ anesthetic/__init__.py
@@ -2,7 +2,7 @@
 import anesthetic.samples
 import anesthetic.plot
 import anesthetic.read.chain
-import anesthetic.io
+import anesthetic.read.hdf
 
 import pandas
 import pandas.plotting._core
@@ -48,5 +48,5 @@ NestedSamples = anesthetic.samples.NestedSamples
 make_2d_axes = anesthetic.plot.make_2d_axes
 make_1d_axes = anesthetic.plot.make_1d_axes
 
-read_hdf = anesthetic.io.read_hdf
+read_hdf = anesthetic.read.hdf.read_hdf
 read_chains = anesthetic.read.chain.read_chains
